requirements.txt
fastapi==0.104.1
hypercorn==0.15.0
openai-whisper==20231117
pydub==0.25.1
aiohttp==3.9.1
numpy<2.0
python-dotenv==1.0.0

